new feature in setter and getter
let have category and product class. product has category as imbricated object

before to hydrate the product class we did:

$product = new Product();
$product->description = 'description';
$product->category = new Category(1); // or may be any category instance
$product->__insert();

the point here is the way we set the category attribut. we were obligated to set an instance of category object and it could be sometime anoying if we just have the category id.

we have the plesure to annonce you that it has been solved. now we can do

....
$product->category_id = $idcategory;
...

like that and devups will automatically do the rest :D

other point was when we create a new instance of product as we did before the attribut $product->category_id was not directly available from now it's ok

its now possible to use global action such as multiple delete and export data