fix issue on migration : unable to add or remove foreign key while migration with devups/doctrine
often when we are developping a project it happen to delete data without taking in consideration the contrain of foreign key or maybe some text anyway something happen for the data persisted to not be well linked with foreign key

therefore when we want to migrate the schema of the database we have the issue of  unable to add or remove foreign key because, while testing the foreign key mysql detect unconsistent relation then it throw the exception.

to fix it we have to sometime drop the entire database beause going through hundred of table is really ennoying  but the problem is that we lose time and test data.

we the fine how to sanitize the unconsistent data within the database while migrating. therefore no more issue of migration due to unconsistent data in dev environment and even more report of  entity and unconsistent data that have been sanitized.